HSEARCH-4566 Move SavedState to a more appropriate package
